新增、修改宠物记录

+++ backend/src/main/java/com/tailYY/backend/model/Pet.java
@@ -31,7 +31,7 @@ public class Pet implements Serializable {
     /**
      * 宠物种类
      */
-    private Integer classId;
+    private Long classId;
 
     /**
      * 宠物信息
